Added code to upload complete PDF file to blob storage

+++ app/backend/Services/AzureBlobStorageService.cs
@@ -38,34 +38,62 @@ internal sealed class AzureBlobStorageService(BlobContainerClient container)
                 else if (Path.GetExtension(fileName).ToLower() is ".pdf")
                 {
                     using var documents = PdfReader.Open(stream, PdfDocumentOpenMode.Import);
-                    for (int i = 0; i < documents.PageCount; i++)
+
+                    //Adding code to also upload original PDF so we have the entire document for reference.
+                    //START NEW CODE FOR UPLOADING ORIGINAL DOCUMENT
+
+                    if (documents.PageCount == 1)
                     {
-                        var documentName = BlobNameFromFilePage(fileName, i);
+                        var blobName = BlobNameFromFilePage(fileName);
+                        var documentName = BlobNameFromFilePage(fileName);
                         var blobClient = container.GetBlobClient(documentName);
+
                         if (await blobClient.ExistsAsync(cancellationToken))
                         {
                             continue;
                         }
 
-                        var tempFileName = Path.GetTempFileName();
-
-                        try
+                        var url = blobClient.Uri.AbsoluteUri;
+                        await using var fileStream = file.OpenReadStream();
+                        await blobClient.UploadAsync(fileStream, new BlobHttpHeaders
                         {
-                            using var document = new PdfDocument();
-                            document.AddPage(documents.Pages[i]);
-                            document.Save(tempFileName);
+                            ContentType = "image"
+                        }, cancellationToken: cancellationToken);
+                        uploadedFiles.Add(blobName);
 
-                            await using var tempStream = File.OpenRead(tempFileName);
-                            await blobClient.UploadAsync(tempStream, new BlobHttpHeaders
+                    }  //END NEW CODE FOR UPLOADING ORIGINAL DOCUMENT
+                    else
+                    {
+
+                        for (int i = 0; i < documents.PageCount; i++)
+                        {
+                            var documentName = BlobNameFromFilePage(fileName, i);
+                            var blobClient = container.GetBlobClient(documentName);
+                            if (await blobClient.ExistsAsync(cancellationToken))
                             {
-                                ContentType = "application/pdf"
-                            }, cancellationToken: cancellationToken);
+                                continue;
+                            }
 
-                            uploadedFiles.Add(documentName);
-                        }
-                        finally
-                        {
-                            File.Delete(tempFileName);
+                            var tempFileName = Path.GetTempFileName();
+
+                            try
+                            {
+                                using var document = new PdfDocument();
+                                document.AddPage(documents.Pages[i]);
+                                document.Save(tempFileName);
+
+                                await using var tempStream = File.OpenRead(tempFileName);
+                                await blobClient.UploadAsync(tempStream, new BlobHttpHeaders
+                                {
+                                    ContentType = "application/pdf"
+                                }, cancellationToken: cancellationToken);
+
+                                uploadedFiles.Add(documentName);
+                            }
+                            finally
+                            {
+                                File.Delete(tempFileName);
+                            }
                         }
                     }
                 }
